requirements.txt
# NOTE: FastAPI < 0.100.0 and Pydantic v1 is required until langchain has full pydantic v2 compatibility
# https://python.langchain.com/v0.1/docs/guides/development/pydantic_compatibility/
# https://github.com/langchain-ai/langchain/discussions/24405
# https://github.com/langchain-ai/langchain/discussions/9337
# IMPORTANT: This also requires using python < 3.12.4

aiohttp~=3.10.0
aiosqlite~=0.20.0
arxiv~=2.1.3
duckduckgo-search~=6.2.6
fastapi<0.100.0
langchain-core~=0.2.26
langchain-community~=0.2.11
langchain-openai~=0.1.20
langchain-groq~=0.1.9
chromadb~=0.5.5
pypdf~=4.3.1
langgraph==0.1.19
langsmith~=0.1.96
numexpr~=2.10.1
pydantic~=1.10.17
python-dotenv~=1.0.1
psycopg[binary]
streamlit~=1.37.0
uvicorn~=0.30.5
langgraph-checkpoint-postgres
